Change content length limit implementation to stream based for MultiSearch operation
Created SizeLimitedCountingInputStream which extends CountingInputStream, to be used when reading InputStream of HTTP response body when we want to abort reading the stream if it is too long. This is done because the HTTP entity in the response does not provide an API to get the content length in some cases.

In order to use this, we need to manually invoke `gson.fromJson` and construct the MultiSearchResult object since the existing implementation just fetches the entire body during deserialization. The rest of the new code in JestHttpClient is completely identical to the regular implementation.

Also had to change the `commons-io` dependency scope to `compile` since it's declared in the parent as `test`